% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine.vcf.slot}
\alias{combine.vcf.slot}
\title{combine different slots of VCF object}
\usage{
combine.vcf.slot(myvcf)
}
\arguments{
\item{myvcf}{a VCF object}
}
\value{
a data.frame
}
\description{
combine different slots of VCF object
}
