% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{download.annovar.db}
\alias{download.annovar.db}
\title{download ANNOVAR DB for annotation
https://doc-openbio.readthedocs.io/projects/annovar/en/latest/user-guide/download/#additional-databases}
\usage{
download.annovar.db(
  force = FALSE,
  path_to_annovar = NA,
  path_annovar_db = "DB_annovar"
)
}
\arguments{
\item{force}{a boolean wheter the function should check if the database is already stored before to download them}

\item{path_to_annovar}{define path to ANNOVAR binaries}

\item{path_annovar_db}{a path where the database will be stored}
}
\description{
download ANNOVAR DB for annotation
https://doc-openbio.readthedocs.io/projects/annovar/en/latest/user-guide/download/#additional-databases
}
