% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.morfee.data}
\alias{get.morfee.data}
\title{get data used by MORFEE}
\usage{
get.morfee.data(path = tempdir(), force = FALSE)
}
\arguments{
\item{path}{is a path to the temporary folder to store downloaded databases}

\item{force}{a boolean wheter the function should check if the database is already stored before to download them}
}
\description{
get data used by MORFEE
}
