% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{is.annovar.installed}
\alias{is.annovar.installed}
\title{check if ANNOVAR is installed
otherwise open the website of ANNOVAR}
\usage{
is.annovar.installed(path_to_annovar = NULL, open_browser = FALSE)
}
\arguments{
\item{path_to_annovar}{define path to ANNOVAR binaries}

\item{open_browser}{booleen if TRUE, when ANNOVAR is not detected a web-browser is open to the ANNOVAR's website}
}
\description{
check if ANNOVAR is installed
otherwise open the website of ANNOVAR
}
