% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morfee.R
\name{morfee.annotation}
\alias{morfee.annotation}
\title{vcf annotation with MORFEE}
\usage{
morfee.annotation(myvcf_annot, morfee_data)
}
\arguments{
\item{myvcf_annot}{an ANNOVAR annotated VCF object to annotate with MORFEE}

\item{morfee_data}{data object obtained from get.morfee.data()}
}
\description{
vcf annotation with MORFEE
}
