% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_GeneDetail.refGene}
\alias{parse_GeneDetail.refGene}
\title{parse refGene}
\usage{
parse_GeneDetail.refGene(x)
}
\arguments{
\item{x}{a string refGene mutation: "NM_001010939:c.-94G>A"}
}
\value{
a vector of different element of the mutation: c("NM_001010939","c.-94G>A","-","94","G","A")
}
\description{
parse refGene
}
