% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_mutation}
\alias{parse_mutation}
\title{parse mutation}
\usage{
parse_mutation(x)
}
\arguments{
\item{x}{a string mutation: "-94G>A"}
}
\value{
a vector of different element of the mutation: c("-","94","G","A")
}
\description{
parse mutation
}
