% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{vcf.annovar.annotation}
\alias{vcf.annovar.annotation}
\title{vcf annotation with ANNOVAR}
\usage{
vcf.annovar.annotation(
  my_raw_vcf,
  path_results = "MORFEE_results",
  path_annovar_db = "DB_annovar"
)
}
\arguments{
\item{my_raw_vcf}{a string to define name of raw VCF to annotate}

\item{path_results}{a path where the annoated VCF will be saved}

\item{path_annovar_db}{a path where the database will be stored}
}
\description{
vcf annotation with ANNOVAR
}
