% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vcf_2_ods}
\alias{vcf_2_ods}
\title{write a VCF object as an .ods OpenDocument file}
\usage{
vcf_2_ods(myvcf, file)
}
\arguments{
\item{myvcf}{a VCF object}

\item{file}{name of the file to be created}
}
\description{
write a VCF object as an .ods OpenDocument file
}
