% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vcf_2_xlsx}
\alias{vcf_2_xlsx}
\title{write a VCF object as an .xlsx Office Open XML file}
\usage{
vcf_2_xlsx(myvcf, file)
}
\arguments{
\item{myvcf}{a VCF object}

\item{file}{name of the file to be created}
}
\description{
write a VCF object as an .xlsx Office Open XML file
}
